requirements.txt
# PyTorch and related packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Data Science and Numerical Computing
numpy>=1.24.0
matplotlib>=3.7.0

# Progress bars and logging
tqdm>=4.65.0

# Configuration files
PyYAML>=6.0

# Additional packages for inference script
pathlib2>=2.3.7;python_version<"3.4"  # Path handling (built-in for Python 3.4+)
argparse>=1.4.0;python_version<"2.7"  # Command line parsing (built-in for Python 2.7+)

# Video processing and webcam support
imageio>=2.31.0          # Video I/O operations
imageio-ffmpeg>=0.4.8    # FFmpeg wrapper for video processing

# Optional: For better performance and additional features
# Uncomment if needed:
# albumentations>=1.3.0  # Advanced image augmentations
# tensorboard>=2.13.0    # Training visualization
# wandb>=0.15.0          # Experiment tracking
# seaborn>=0.12.0        # Enhanced plotting

# Optional: Enhanced video processing
# ffmpeg-python>=0.2.0   # More advanced FFmpeg operations
# scikit-image>=0.21.0   # Additional image processing utilities

# Development tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0